make sure the scheduled updates are always performed
with the previous throttling we was completely ignoring some updates even if they was manually fired by user. Now we make sure to schedule a new update ahead in time to make sure the update is always correctly executed.
In pratice, we delay the updates by 16ms instead of dropping them. This keeps the performance improvements and doesn't affect the behavior.