Calculate boundaries relative to position (Fixes #96)
When a specific boundaryElement is specified, the position must be
calculated relative to the popper.

In a fixed-position popper, this is the offset rectangle outside of the
boundary element.
In an absolutely positioned popper, this is the width of the scroll
offset element, minus the left.

This patch also corrects the padding on the bottom and right.