Remove event listeners properly on .dispose()
These event listeners are added at Tooltip construction time, whereas `_tooltipNode` is added the first time the tooltip is shown. That means the listeners will always be added on the target node, whether we end up showing the tooltip or not. We should then remove them regardless of the existence of `_tooltipNode`