Remove /src from .npmignore (for debugging)
Popper publishes source maps to NPM (great!) but it doesn't publish the /src folder that the sourcemaps point to. This makes it harder to debug into & set breakpoints in popper code from Chrome or an IDE like VSCode.  This PR updates .npmignore to ensure that the /src folder is also, along with /dist, is published to NPM.