Fix behavior for wrapped buttons
The current implementation of `useClick` tries to bail on keyDown/Up handlers on button targets to avoid redundant handling of popover state because pressing ENTER/SPACE on buttons trigger a click event. Checking if the reference element is a button causes it to incorrectly bail on the wrong element type. Instead, one possible solution is to check if the target that triggered the event is a button regardless of the type of the wrapping reference.

[Here](https://codesandbox.io/s/serene-noether-bqk39m?file=/src/App.tsx)'s a reproduction of the issue.